% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSetCalculation.R
\name{FeatureSetCalculationComponent}
\alias{FeatureSetCalculationComponent}
\title{Shiny component for Calculating and displaying flacco FeatureSets}
\usage{
FeatureSetCalculationComponent(id)
}
\arguments{
\item{id}{ID for the shiny component}
}
\description{
\code{FeatureSetCalculationComponent} is a shiny component which can be added to your shiny app
so that you can calculate and display different Feature Sets.
}
\details{
It will get a flacco feature.object for the next steps
}
