% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSetVisualization.R
\name{FeatureSetVisualization}
\alias{FeatureSetVisualization}
\title{Shiny server function for FeatureSet Component}
\usage{
FeatureSetVisualization(input, output, session, stringsAsFactors, feat.object)
}
\arguments{
\item{input}{Shiny input variable for the specific UI module}

\item{output}{Shiny output variable for the specific UI module}

\item{session}{Shiny session variable for the specific UI module}

\item{stringAsFactors}{}
}
\description{
\code{FeatureSetVisualization} is a Shiny server function which will control all aspects
of the FeatureSetVisualizationComponent UI Module. Will be called with \code{callModule()}
}
