% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSetVisualization.R
\name{FeatureSetVisualizationComponent}
\alias{FeatureSetVisualizationComponent}
\title{Shiny component for visualizing flacco FeatureSets}
\usage{
FeatureSetVisualizationComponent(id)
}
\arguments{
\item{id}{ID for the shiny component}
}
\description{
\code{FeatureSetVisualizationComponent} is a shiny component which can be added to your shiny app
so that you can display different Feature Set Graphs.
}
\details{
It will get a flacco feature.object for the next steps
}
