% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoofFunction.R
\name{SmoofFunctionInput}
\alias{SmoofFunctionInput}
\title{Shiny server function for BBOB import page module}
\usage{
SmoofFunctionInput(input, output, session, stringsAsFactors)
}
\arguments{
\item{input}{Shiny input variable for the specific UI module}

\item{output}{Shiny output variable for the specific UI module}

\item{session}{Shiny session variable for the specific UI module}

\item{stringAsFactors}{}
}
\description{
\code{SmoofFunctionInput} is a Shiny server function which will control all aspects
of the SmoofFunctionPage UI Module. Will be called with \code{callModule()}
}
