% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoofFunction.R
\name{SmoofFunctionPage}
\alias{SmoofFunctionPage}
\title{Shiny component for Smoof-Function Input}
\usage{
SmoofFunctionPage(id)
}
\arguments{
\item{id}{ID for the shiny component}
}
\description{
\code{SmoofFunctionPage} is a shiny component which can be added to your shiny app
so that you can put in a smoof function and calculate flacco features after that
}
